Refactor component structure: Separate image and navigation components
-Created a new component for image handling.
-Refactored navigation to its own separate component.
-Improved the modularity and maintainability of the code by separating concerns.
