+++ src/app/login/login-page.module.css
@@ -1,7 +1,6 @@
-
 .page_wrapper {
   display: flex;
-  width: 100%;
+  width: 100vw;
   height: 100vh;
   background-color: var(--bg-color);
   color: var(--text-color);
@@ -10,65 +9,51 @@
 .login_form_wrapper {
   display: flex;
   flex-direction: column;
-  align-items: center;
   justify-content: flex-start;
-  width: 50%;
+  align-items: center;
+  width: 50%; /* Adjust width to make room for the image */
   height: 100%;
   padding: 20px;
 }
-.login_form_wrapper h3{
-  font-size: xx-large;
-  font-weight: 500;
-}
 
 .login_image {
-  max-width: 100%;
-  max-height: 100%;
-  border-radius: .5rem;
-  object-fit: contain;
-  margin-bottom: 20px;
   display: flex;
   justify-content: center;
   align-items: center;
+  width: 50%; /* Adjust width for image placement */
+  height: 100%;
+  border-radius: 0.5rem;
+  object-fit: contain;
+  margin: 0; /* Ensure no unwanted spacing */
 }
 
 .logo_wrapper {
   width: 100%;
+  margin-bottom: 1rem;
 }
-@media only screen and (max-width: 767px) 
-{ .login_form_wrapper 
-  { width: 100%; padding: 20px; 
-  } 
-  .login_image{
+
+@media only screen and (max-width: 767px) {
+  .login_form_wrapper {
+    width: 100%;
+    padding: 20px;
+  }
+  .login_image {
     display: none;
   }
-  .login_form_wrapper h3{
-    font-size: x-large;
-    font-weight: 400;
-  }
 }
+
 @media (min-width: 768px) and (max-width: 1024px) {
   .login_form_wrapper {
     width: 100%;
-    padding: 20px; 
-    align-items: center;
+    padding: 20px;
   }
-
   .login_image {
     display: none;
   }
-
-  .page_wrapper {
-    width: 100%;
-  }
 }
+
 @media only screen and (min-width: 1025px) {
   .login_image {
-      display: block;
-      align-items: center;
-  }
-
-  .login_image>div {
-      margin-top: 5rem;
+    display: flex;
   }
 }
