+++ src/components/login/image.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import Image from "next/image";
+
+export const LogoImage = () => (
+  <Image src="/orbglo-logo.png" alt="logo" width="100" height="100" />
+);
+
+export const LoginImage = () => (
+  <Image src="/login.png" alt="login illustration" width="800" height="500" />
+);
