+++ src/components/login/nav.tsx
@@ -1,0 +1,15 @@
+
+import React from "react";
+import { IconButton } from "@mui/joy";
+import { DarkMode, LightMode } from "@mui/icons-material";
+import { ThemeContext } from "@/context/ThemeContext";
+
+export const NavBar = () => {
+  const { mode, handleTheme } = React.useContext(ThemeContext);
+
+  return (
+    <IconButton onClick={handleTheme} sx={{ position: "absolute", top: 16, right: 16 }}>
+      {mode === "dark" ? <DarkMode /> : <LightMode />}
+    </IconButton>
+  );
+};
